# openapi/swagger.yaml
basePath: /
definitions:
...
  github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorBody'
    type: object
...
  internal_metering_adapters_http_usage.EventListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_metering_adapters_http_usage.Response'
        type: array
      next_cursor:
        type: string
    type: object
  internal_metering_adapters_http_usage.EventSearchRequest:
    properties:
      cursor:
        type: string
      filter:
        $ref: '#/definitions/internal_metering_adapters_http_usage.FilterRequest'
      from:
        type: string
      limit:
        type: integer
      meter:
        type: string
      subject:
        type: string
      to:
        type: string
    type: object
  internal_metering_adapters_http_usage.ExportJobCreateRequest:
    properties:
      format:
        type: string
      kind:
        type: string
      query:
        $ref: '#/definitions/internal_metering_adapters_http_usage.SearchRequest'
    type: object
  internal_metering_adapters_http_usage.ExportJobListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_metering_adapters_http_usage.ExportJobResponse'
        type: array
      next_cursor:
        type: string
    type: object
  internal_metering_adapters_http_usage.ExportJobResponse:
    properties:
      artifact_size:
        type: integer
      completed_at:
        type: string
      created_at:
        type: string
      download_url:
        type: string
      error:
        type: string
      format:
        type: string
      id:
        type: string
      kind:
        type: string
      query:
        $ref: '#/definitions/internal_metering_adapters_http_usage.SearchRequest'
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  internal_metering_adapters_http_usage.PruneListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_metering_adapters_http_usage.PruneResponse'
        type: array
      next_cursor:
        type: string
    type: object
...
  internal_metering_adapters_http_usage.PruneResponse:
    properties:
      created_at:
        type: string
      deleted:
        type: integer
      dry_run:
        type: boolean
      id:
        type: string
      meters:
        items:
          $ref: '#/definitions/internal_metering_adapters_http_usage.PruneMeterResponse'
        type: array
    type: object
...
info:
  contact: {}
  description: Open source metering service for tracking who used what, when, how
    much, and in what context.
  title: Open Spanner API
  version: 0.1.3
paths:
...
  /v1/exports:
    get:
      operationId: listUsageExportJobs
      parameters:
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_metering_adapters_http_usage.ExportJobListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
      summary: List usage export jobs
      tags:
      - exports
    post:
      consumes:
      - application/json
      description: Queues a usage bucket export job. Jobs are created as queued and
        can be listed or inspected by ID.
      operationId: createUsageExportJob
      parameters:
      - description: Usage export job
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_metering_adapters_http_usage.ExportJobCreateRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/internal_metering_adapters_http_usage.ExportJobResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
      summary: Queue usage export
      tags:
      - exports
  /v1/exports/{id}:
    get:
      operationId: getUsageExportJob
      parameters:
      - description: Export job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_metering_adapters_http_usage.ExportJobResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
      summary: Get usage export job
      tags:
      - exports
  /v1/exports/{id}/cancel:
    post:
      operationId: cancelUsageExportJob
      parameters:
      - description: Export job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_metering_adapters_http_usage.ExportJobResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
      summary: Cancel usage export job
      tags:
      - exports
  /v1/exports/{id}/download:
    get:
      operationId: downloadUsageExportJob
      parameters:
      - description: Export job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/csv
      responses:
        "200":
          description: CSV
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
      summary: Download usage export job
      tags:
      - exports
  /v1/exports/{id}/retry:
    post:
      operationId: retryUsageExportJob
      parameters:
      - description: Export job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_metering_adapters_http_usage.ExportJobResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
      summary: Retry usage export job
      tags:
      - exports
...
  /v1/usageevents:
    get:
      operationId: listUsageEvents
      parameters:
      - description: Subject
        in: query
        name: subject
        type: string
      - description: Meter name
        in: query
        name: meter
        type: string
      - description: RFC3339 start time
        in: query
        name: from
        type: string
      - description: RFC3339 end time
        in: query
        name: to
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_metering_adapters_http_usage.EventListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
      summary: List usage events
      tags:
      - usage-events
  /v1/usageevents/export:
    get:
      operationId: exportUsageEvents
      parameters:
      - description: Subject
        in: query
        name: subject
        type: string
      - description: Meter name
        in: query
        name: meter
        type: string
      - description: RFC3339 start time
        in: query
        name: from
        type: string
      - description: RFC3339 end time
        in: query
        name: to
        type: string
      - description: Result limit. Defaults to 100 and cannot exceed 1000.
        in: query
        name: limit
        type: integer
      produces:
      - text/csv
      responses:
        "200":
          description: CSV
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
      summary: Export usage events
      tags:
      - usage-events
    post:
      consumes:
      - application/json
      operationId: exportFilteredUsageEvents
      parameters:
      - description: Usage event export search
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_metering_adapters_http_usage.EventSearchRequest'
      produces:
      - text/csv
      responses:
        "200":
          description: CSV
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
      summary: Export filtered usage events
      tags:
      - usage-events
  /v1/usageevents/prune:
    post:
      operationId: pruneUsageEvents
      parameters:
      - description: Count prunable events without deleting
        in: query
        name: dry_run
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_metering_adapters_http_usage.PruneResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
      summary: Prune usage events
      tags:
      - usage-events
  /v1/usageevents/prunes:
    get:
      operationId: listPruneRuns
      parameters:
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_metering_adapters_http_usage.PruneListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
      summary: List prune runs
      tags:
      - usage-events
  /v1/usageevents/search:
    post:
      consumes:
      - application/json
      operationId: searchUsageEvents
      parameters:
      - description: Usage event search
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_metering_adapters_http_usage.EventSearchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_metering_adapters_http_usage.EventListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
      summary: Search usage events
      tags:
      - usage-events
  /v1/usageingestions:
    get:
      operationId: listIngestionRuns
      parameters:
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_metering_adapters_http_usage.IngestionListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
      summary: List ingestion runs
      tags:
      - usages
  /v1/usages:
    get:
      operationId: listUsageBuckets
      parameters:
      - description: Subject
        in: query
        name: subject
        type: string
      - description: Meter name
        in: query
        name: meter
        required: true
        type: string
      - description: RFC3339 start time
        in: query
        name: from
        required: true
        type: string
      - description: RFC3339 end time
        in: query
        name: to
        required: true
        type: string
      - description: 'Bucket size: hour, day, month'
        in: query
        name: bucket_size
        type: string
      - collectionFormat: multi
        description: Subject or metadata keys to group by. Repeat the parameter or
          use comma-separated values.
        in: query
        items:
          type: string
        name: group_by
        type: array
      - description: Result limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_metering_adapters_http_usage.ListItemResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
      summary: List usage buckets
      tags:
      - usages
    post:
      consumes:
      - application/json
      description: Records one usage event. If idempotency_key matches a previously
        accepted event, the original event is returned. A duplicate event ID is a
        conflict.
      operationId: createUsage
      parameters:
      - description: Usage event
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_metering_adapters_http_usage.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_metering_adapters_http_usage.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
      summary: Create usage
      tags:
      - usages
  /v1/usages/breakdowns/search:
    post:
      consumes:
      - application/json
      operationId: searchUsageBreakdown
      parameters:
      - description: Usage breakdown search
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_metering_adapters_http_usage.BreakdownRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_metering_adapters_http_usage.BreakdownListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
      summary: Search usage breakdown
      tags:
      - usages
  /v1/usages/bulk:
    post:
      consumes:
      - application/json
      description: Records up to 1000 usage events. The Idempotency-Key header replays
        the original bulk response for the same batch. Per-event idempotency_key values
        replay existing events as duplicates. Duplicate event IDs are conflicts.
      operationId: createUsageBulk
      parameters:
      - description: Batch idempotency key. Reusing it returns the original bulk response.
        in: header
        name: Idempotency-Key
        type: string
      - description: Usage events. Maximum 1000 items.
        in: body
        name: request
        required: true
        schema:
          items:
            $ref: '#/definitions/internal_metering_adapters_http_usage.CreateRequest'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_metering_adapters_http_usage.BulkResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
      summary: Create usage in bulk
      tags:
      - usages
  /v1/usages/dimensions:
    get:
      operationId: listUsageDimensionValues
      parameters:
      - description: Meter name
        in: query
        name: meter
        required: true
        type: string
      - description: Metadata dimension field
        in: query
        name: field
        required: true
        type: string
      - description: Subject
        in: query
        name: subject
        type: string
      - description: RFC3339 start time
        in: query
        name: from
        type: string
      - description: RFC3339 end time
        in: query
        name: to
        type: string
      - description: Result limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_metering_adapters_http_usage.DimensionValueListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
      summary: List usage dimension values
      tags:
      - usages
  /v1/usages/export:
    get:
      operationId: exportUsageBuckets
      parameters:
      - description: Subject
        in: query
        name: subject
        type: string
      - description: Meter name
        in: query
        name: meter
        required: true
        type: string
      - description: RFC3339 start time
        in: query
        name: from
        required: true
        type: string
      - description: RFC3339 end time
        in: query
        name: to
        required: true
        type: string
      - description: 'Bucket size: hour, day, month'
        in: query
        name: bucket_size
        type: string
      - collectionFormat: multi
        description: Subject or metadata keys to group by. Repeat the parameter or
          use comma-separated values.
        in: query
        items:
          type: string
        name: group_by
        type: array
      - description: Result limit. Defaults to 100 and cannot exceed 1000.
        in: query
        name: limit
        type: integer
      produces:
      - text/csv
      responses:
        "200":
          description: CSV
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
      summary: Export usage buckets
      tags:
      - usages
    post:
      consumes:
      - application/json
      operationId: exportFilteredUsageBuckets
      parameters:
      - description: Usage export search
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_metering_adapters_http_usage.SearchRequest'
      produces:
      - text/csv
      responses:
        "200":
          description: CSV
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ssubedir_open-spanner_internal_metering_adapters_http_internal_respond.ErrorResponse'
      summary: Export filtered usage buckets
      tags:
      - usages
...
schemes:
- http
swagger: "2.0"